requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.4.2
pandas==2.1.3
numpy==1.26.2
python-dotenv==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.24.1 